+++ NEWS.md
@@ -1,3 +1,20 @@
+# sjPlot 2.8.2
+
+## Function renaming
+
+* `sjt.itemanalysis()` is now named `tab_itemscale()`.
+* `sjt.xtab()` is now named `tab_xtab()`.
+
+## Changes to functions
+
+* Improved handling for `tab_model()` of robust estimation in general and Kenward-Roger or Satterthwaite approximations in particular for linear mixed models.
+* Revised code to cope with forthcoming *tidyselect*-update.
+
+## Bug fixes
+
+* Improved `tab_df()` now uses value labels for factors instead of numeric values.
+* Fixed some issues related to the lates *brms*-update.
+
 # sjPlot 2.8.1
 
 ## Changes to functions